[QUYEN][QUEENSTORE][PRODUCT] Create model and repo
