+++ queen_store/src/main/java/com/example/queen_store/common/product/ProductCommon.java
@@ -1,4 +1,4 @@
 package com.example.queen_store.common.product;
 
-public class Product {
+public class ProductCommon {
 }
