+++ queen_store/src/main/java/com/example/queen_store/controller/product/ProductServlet.java
@@ -1,0 +1,19 @@
+package com.example.queen_store.controller.product;
+
+import javax.servlet.*;
+import javax.servlet.http.*;
+import javax.servlet.annotation.*;
+import java.io.IOException;
+
+@WebServlet(name = "ProductServlet", value = "/ProductServlet")
+public class ProductServlet extends HttpServlet {
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
+    }
+}
