+++ queen_store/src/main/java/com/example/queen_store/model/product/Product.java
@@ -1,4 +1,60 @@
 package com.example.queen_store.model.product;
 
 public class Product {
+    private int id;
+    private String name;
+    private double price;
+    private String description;
+    private String type;
+
+    public Product() {
+    }
+
+    public Product(int id, String name, double price, String description, String type) {
+        this.id = id;
+        this.name = name;
+        this.price = price;
+        this.description = description;
+        this.type = type;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public double getPrice() {
+        return price;
+    }
+
+    public void setPrice(double price) {
+        this.price = price;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
 }
