+++ queen_store/src/main/java/com/example/queen_store/repository/product/IProductRepository.java
@@ -1,2 +1,9 @@
-package com.example.queen_store.repository.product;public interface IProductRepository {
+package com.example.queen_store.repository.product;
+
+import com.example.queen_store.model.product.Product;
+
+import java.util.List;
+
+public interface IProductRepository {
+    List<Product> getAll();
 }
