+++ queen_store/src/main/java/com/example/queen_store/service/product/IProductService.java
@@ -1,2 +1,4 @@
-package com.example.queen_store.service.product;public interface IProductService {
+package com.example.queen_store.service.product;
+
+public interface IProductService {
 }
