+++ queen_store/src/main/java/com/example/queen_store/service/product/ProductService.java
@@ -1,4 +1,4 @@
 package com.example.queen_store.service.product;
 
-public class Product {
+public class ProductService implements IProductService {
 }
